server/requirements.txt
# === CORE ASGI & API ===
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.12

# === LLM ORCHESTRATION (LangGraph + LangChain) ===
langgraph==0.3.5
langchain==0.3.26
langchain-core==0.3.67
langchain-community==0.3.27

# === GOOGLE VERTEX AI (DOWNGRADED TO COMPATIBLE VERSION) ===
langchain-google-vertexai==2.0.4 

# === LLM BACKEND ===
groq==0.11.0

# === GOOGLE CLOUD ===
google-cloud-aiplatform==1.70.0
google-cloud-translate==3.15.5
google-auth==2.35.0

# === VECTOR DB & EMBEDDINGS ===
faiss-cpu==1.8.0
numpy==1.26.4  # ← CRITICAL: No 2.0 breakage

# === PDF GENERATION ===
weasyprint==66.0
reportlab==4.2.5

# === WHATSAPP / TWILIO ===
twilio==9.3.0

# === AUTH & SECURITY ===
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# === OSM & UTILS ===
overpy==0.7
httpx==0.27.2
websockets==13.1

# === DATA (safe with NumPy 1.26.4) ===
pandas==2.2.3
scipy==1.13.1

# === MISC ===
pydantic==2.9.2
python-jose[cryptography]==3.3.0